This fixes #84 and implements the hashing of passwords in the database.
Update .gitignore

Add .idea to gitignore.

Add hashing of password

Implemented password-hash for regestration, so that the password is not saved in clear in the database. Also the password is now checked against the hash in the database when logging in.

Update docs

Update version.go

Reomve unnecessary imports

Improved error-handling

Add id/_id to the response

Added id/_id to the response that is returned to the calling party, so that it knows the id of the logged in user.

Remove name from token

Remove name from token when loggin in

Update modules/userman/email.go

Only filter for email when looking if an user already exists.

Co-Authored-By: ledex <karstenbruhns@gmail.com>